[API,CLI](CIVO) merging the civo/ and ha_civo/ folders
