--- api/ha_civo/main_test.go
@@ -1,12 +1,0 @@
-package ha_civo
-
-import (
-	"fmt"
-	"testing"
-)
-
-// TODO: ADD TEST CASES
-
-func TestGenerateDBPassword(t *testing.T) {
-	fmt.Println(generateDBPassword(20))
-}
