+++ cli/cmd/createClusterHACivo.go
@@ -9,7 +9,8 @@ Kubesimplify
 import (
 	"fmt"
 
-	"github.com/kubesimplify/ksctl/api/ha_civo"
+	"github.com/kubesimplify/ksctl/api/civo"
+	"github.com/kubesimplify/ksctl/api/utils"
 	"github.com/spf13/cobra"
 )
 
@@ -21,7 +22,25 @@ var createClusterHACivo = &cobra.Command{
 ksctl create-cluster ha-civo <arguments to civo cloud provider>
 `,
 	Run: func(cmd *cobra.Command, args []string) {
-		err := ha_civo.CreateCluster(chcclustername, chcregion, chcnodesize, chcnocp, chcnowp)
+		payload := civo.CivoProvider{
+			ClusterName: chcclustername,
+			Region:      chcregion,
+			HACluster:   true,
+			Spec: utils.Machine{
+				Disk:                chcnodesize,
+				HAControlPlaneNodes: chcnocp,
+				HAWorkerNodes:       chcnowp,
+			},
+		}
+		// clusterConfig := civoHandler.ClusterInfoInjecter(
+		// 	cclusterName,
+		// 	cregion,
+		// 	cspec.Disk,
+		// 	cspec.Nodes,
+		// 	apps,
+		// 	cni)
+		err := payload.CreateCluster()
+		// err := ha_civo.CreateCluster(chcclustername, chcregion, chcnodesize, chcnocp, chcnowp)
 		if err != nil {
 			fmt.Printf("\033[31;40m%v\033[0m\n", err)
 		}
