+++ cli/cmd/createClusterLocal.go
@@ -23,7 +23,7 @@ var createClusterLocal = &cobra.Command{
 ksctl create-cluster local <arguments to civo cloud provider>
 `,
 	Run: func(cmd *cobra.Command, args []string) {
-		cargo := local.ClusterInfoInjecter(clocalclusterName, clocalspec.Nodes)
+		cargo := local.ClusterInfoInjecter(clocalclusterName, clocalspec.ManagedNodes)
 		fmt.Println("Building...")
 		if err := local.CreateCluster(cargo); err != nil {
 			fmt.Printf("\033[31;40m%v\033[0m\n", err)
@@ -48,7 +48,7 @@ var (
 func init() {
 	createClusterCmd.AddCommand(createClusterLocal)
 	createClusterLocal.Flags().StringVarP(&clocalclusterName, "name", "n", "demo", "Cluster name")
-	createClusterLocal.Flags().IntVarP(&clocalspec.Nodes, "nodes", "N", 1, "Number of Nodes")
+	createClusterLocal.Flags().IntVarP(&clocalspec.ManagedNodes, "nodes", "N", 1, "Number of Nodes")
 	createClusterLocal.MarkFlagRequired("name")
 	//createClusterLocal.MarkFlagRequired("nodes")
 }
